backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pypdf==4.2.0
faiss-cpu==1.8.0
numpy==1.26.4
requests==2.32.3
pydantic==2.7.1
python-dotenv==1.0.1
sentence-transformers==2.7.0
torch==2.2.2
transformers==4.41.2